bridge: fix hotplug vlan overwrite on big-endian systems
The avl key type for bridge vlans is uint16_t, so any lookup with a wider
type is going to fail on big-endian systems
This resulted in hotplug-added devices replacing configured member ports

+++ bridge.c
@@ -690,7 +690,7 @@ bridge_add_member(struct bridge_state *bst, const char *name)
 }
 
 static struct bridge_vlan *
-bridge_hotplug_get_vlan(struct bridge_state *bst, unsigned int vid)
+bridge_hotplug_get_vlan(struct bridge_state *bst, uint16_t vid)
 {
 	struct bridge_vlan *vlan;
